Move storage to another file
